Feature to disable persisting prepared statements into table
Summary:
[cassandra-4.1-simulator-dtest]

**Context:**

After Cassandra 4.1, we see the local `system.prepared_statements` size started to grow unboundedly.

This provides a feature to skip the persist path for prepared statements.

**Note: Without this local table, we'll not be able to warmup the cache at startup. If the prepare request rate is too high, the cache could possibly lock itself if it's waiting for some expensive calculation for the prepared statement size. One possible workaround for this is to setup the initial cache capacity, or switch to entry number-based cache that doesn't need to calculate the size.**

Related thread: https://github.com/ben-manes/caffeine/issues/672

Test Plan: UT

Reviewers: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Reviewed By: O2017 Cassandra Service, #ldap_cassandra-super-users, runtian

Subscribers: runtian, mingc

JIRA Issues: SO-31073

Differential Revision: https://code.uberinternal.com/D14151695

+++ src/java/org/apache/cassandra/db/SystemKeyspace.java
@@ -1848,16 +1848,31 @@ public final class SystemKeyspace
 
     public static void writePreparedStatement(String loggedKeyspace, MD5Digest key, String cql, long timestamp)
     {
-        executeInternal(format("INSERT INTO %s (logged_keyspace, prepared_id, query_string) VALUES (?, ?, ?) USING TIMESTAMP ?",
-                               PreparedStatements.toString()),
-                        loggedKeyspace, key.byteBuffer(), cql, timestamp);
-        logger.debug("stored prepared statement for logged keyspace '{}': '{}'", loggedKeyspace, cql);
+        if (DatabaseDescriptor.getPersistPreparedStatementsEnabled())
+        {
+            executeInternal(format("INSERT INTO %s (logged_keyspace, prepared_id, query_string) VALUES (?, ?, ?) USING TIMESTAMP ?",
+                                   PreparedStatements.toString()),
+                            loggedKeyspace, key.byteBuffer(), cql, timestamp);
+            logger.debug("stored prepared statement for logged keyspace '{}': '{}'", loggedKeyspace, cql);
+        }
+        else
+        {
+            logger.debug("not storing prepared statement for logged keyspace '{}': '{}', persite_prepared_statement_enabled=false",
+                         loggedKeyspace, cql);
+        }
     }
 
     public static void removePreparedStatement(MD5Digest key)
     {
-        executeInternal(format("DELETE FROM %s WHERE prepared_id = ?", PreparedStatements.toString()),
-                        key.byteBuffer());
+        if (DatabaseDescriptor.getPersistPreparedStatementsEnabled())
+        {
+            executeInternal(format("DELETE FROM %s WHERE prepared_id = ?", PreparedStatements.toString()),
+                            key.byteBuffer());
+        }
+        else
+        {
+            logger.debug("not deleting prepared statement for {}", key);
+        }
     }
 
     public static void resetPreparedStatements()
